Added support for retrieval of current reset confirguration
I'm looking to be able to control the reset counter remotely as in some instances I want to set the reset to be much longer. This is the first step in that change.

+++ src/main/java/dev/jarradclark/tflproxy/services/TFLHelper.java
@@ -43,7 +43,7 @@ public class TFLHelper {
         if (seconds < 60) {
             return "Due";
         } else {
-            return String.format("%sm", Math.round(seconds / 60));
+            return String.format("%sm", Math.round((float) (seconds / 60)));
         }
     }
